Moved `write_shared` to its own file
